+++ beam-portal-toolkit/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- shared storage layer for extensions
+
+## [0.2.0] - 2026-06-23
+
+### Added
+- `ExtensionContext` gains new fields, `params`, `location`, `navigate`, `mount`, `config`, and `updateBadge()`
+- Extensions can add badges to nav bar
+- `definePortalExtensionConfig` vite extensions
+
 ## [0.1.10] - 2026-06-03
 
 ### Changed